requirements.txt
adbutils==2.8.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
apkutils2==1.0.0
Appium-Python-Client==4.2.0
attrs==24.2.0
backports.functools-lru-cache==2.0.0
cached-property==1.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cigam==0.0.3
colorama==0.4.6
construct==2.10.70
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
facebook-wda==1.4.9
future==1.0.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
logzero==1.7.0
lxml==5.3.0
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pycparser==2.22
pyelftools==0.31
PySocks==1.7.1
pytest==8.3.3
pytest-env==1.1.5
python-dotenv==1.0.1
pywin32==306
requests==2.32.3
retry==0.9.2
selene==2.0.0rc9
selenium==4.25.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tornado==6.4.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
uiautomator2==3.2.5
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
weditor==0.7.3
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.13.0
